+++ src/resolver.ts
@@ -30,7 +30,19 @@ export interface IMappedPath {
 }
 
 interface IResolverOptions {
-    vPathMap: IVirtualPathMapping[];
+    /**
+     * Virtual path mappings.
+     */
+    virtPathMaps: IVirtualPathMapping[];
+
+    /**
+     * Current working directory.
+     */
+    cwd: string;
+
+    /**
+     * Logger.
+     */
     logger: Logger;
 }
 
@@ -44,12 +56,13 @@ interface IResolverOptions {
  */
 export default function (globs: string[], options: IResolverOptions): IMappedPath[] {
     const log = options.logger;
+    const cwd = options.cwd;
 
     // Resolve all file paths
 
     log.trace("Looking for files", { globs: globs });
 
-    const paths = new Set<string>(globby.sync(globs));
+    const paths = new Set<string>(globby.sync(globs, { cwd }));
 
     log.trace(`Found ${paths.size} files, filtering with provided virtual paths`);
 
@@ -61,28 +74,28 @@ export default function (globs: string[], options: IResolverOptions): IMappedPat
      * - 1 is the real path
      * - 2 is the file perference
      */
-    const candidatePaths = new Map<string, [string, number]>();
+    const candidatePaths = new Map<string, { actual: string, preference: number}>();
 
-    for (const path of paths) {
-        log.trace("Inspecting file path", { path });
+    for (const actual of paths) {
+        log.trace("Inspecting file path", { path: actual });
 
         // Resolve path
-        const [ vPath, preference ] = resolveVirtualPath(path, options.vPathMap, log);
+        const [ virtual, preference ] = resolveVirtualPath(actual, options.virtPathMaps, cwd, log);
 
         // Add to resolved file if highest preference
-        const existingFile = candidatePaths.get(vPath);
+        const existingFile = candidatePaths.get(virtual);
         if (existingFile) {
-            if (existingFile[1] < preference) {
-                log.trace("Superseeding existing file path", { vPath, path, preference });
-                candidatePaths.set(vPath, [ path, preference ]);
+            if (existingFile.preference < preference) {
+                log.trace("Superseeding existing file path", { virtual, actual, preference });
+                candidatePaths.set(virtual, { actual, preference });
             }
             else {
-                log.trace("Discarding new file path due to lower preference", { vPath, path, preference });
+                log.trace("Discarding new file path due to lower preference", { virtual, actual, preference });
             }
         }
         else {
-            log.trace("Adding file path", { vPath, path, preference });
-            candidatePaths.set(vPath, [ path, preference ]);
+            log.trace("Adding file path", { virtual, actual, preference });
+            candidatePaths.set(virtual, { actual, preference });
         }
     }
 
@@ -92,8 +105,8 @@ export default function (globs: string[], options: IResolverOptions): IMappedPat
 
     const resolvedPaths: IMappedPath[] = [];
 
-    for (const [ vPath, [ path ] ] of candidatePaths.entries()) {
-        resolvedPaths.push({ virtual: resolvePath(vPath), actual: resolvePath(path) });
+    for (const [ virtual, { actual } ] of candidatePaths.entries()) {
+        resolvedPaths.push({ virtual, actual: resolvePath(cwd, actual) });
     }
 
     return resolvedPaths;
@@ -102,29 +115,34 @@ export default function (globs: string[], options: IResolverOptions): IMappedPat
 /**
  * Attempts to resolve path to a virtual path, returning provided path on failure.
  *
- * @param path - Absolute path to try and resolve.
- * @param vPaths - Virtual path mappings.
+ * @param actual - Absolute path to try and resolve.
+ * @param virtPathMaps - Virtual path mappings.
  * @param log - Logger
  *
  * @returns New or existing path and preference.
  */
-function resolveVirtualPath(path: string, vPaths: IVirtualPathMapping[], log: Logger): [string, number] {
+function resolveVirtualPath(
+    actual: string,
+    virtPathMaps: IVirtualPathMapping[],
+    cwd: string,
+    log: Logger
+): [string, number] {
     // Try to resolve a virtual path
-    log.trace("Attempting to resolve virtual path", { path });
+    log.trace("Attempting to resolve virtual path", { path: actual });
 
     let preference = 0;
-    for (const { match, replace } of vPaths) {
-        if (path.startsWith(match)) {
+    for (const { match, replace } of virtPathMaps) {
+        if (actual.startsWith(match)) {
             // Virtual path match discovered
-            const resolvedPath = resolvePath(path.replace(match, replace));
-            log.trace("Resolved virtual path", { path, resolvedPath, preference });
-            return [ resolvedPath, preference ];
+            const virtual = resolvePath(cwd, actual.replace(match, replace));
+            log.trace("Resolved virtual path", { actual, virtual, preference });
+            return [ virtual, preference ];
         }
 
         preference++;
     }
 
     // No matches, lowest preference
-    log.trace("No matching virtual path mappings", { path, preference: 0 });
-    return [ path, 0] ;
+    log.trace("No matching virtual path mappings", { actual, preference: 0 });
+    return [ resolvePath(cwd, actual), 0] ;
 }